Fix main-layout
