+++ src/components/footer/footer.module.scss
@@ -2,24 +2,116 @@
 @import "../../css/variables.scss";
 
 .container {
-  justify-content: start;
   display: flex;
+  flex-direction: column;
   overflow: none;
-  height: auto;
+  align-items: start;
+  justify-content: center;
   background-color: $blueOne;
-  flex-direction: row;
   padding: 5rem 0 4rem 15.5%;
-}
 
-@media only screen and (max-width: map-get($map: $screen-breakpoint, $key: xxl)) {
-  .container {
+  @media only screen and (max-width: map-get($map: $screen-breakpoint, $key: xxl)) {
     padding-left: 4%;
   }
+  
+  @media screen and (max-width: map-get($map: $screen-breakpoint, $key: sm)) {
+    height: 100%;
+    padding: 4rem 0 3rem 1.3rem;
+  }
 }
 
-@media screen and (max-width: map-get($map: $screen-breakpoint, $key: sm)) {
-  .container {
-    height: 100%;
-    padding: 4rem 0 3rem 1.3rem;  
+
+.linkContainer {
+  margin-top: 2rem;
+  display: flex;
+  flex-direction: column;
+  color: $defaultGray;
+}
+
+.optionContainer {
+  margin-left: 0.9rem;
+  display: flex;
+  flex-direction: column;
+  align-items: start;
+  justify-content: center;
+
+  a {
+    cursor: pointer;
+    font-family: $primary-font;
   }
+
+  a:hover {
+    color: $neutral100;
+  }
+}
+
+.logo {
+  margin-left: 0.9rem;
+}
+
+
+@media screen and (max-height: map-get($map: $screen-breakpoint, $key: md)) {
+  .filterContainer {
+    flex-wrap: wrap;
+  }
+
+  .filterElement {
+    margin-bottom: 2rem;
+  }
+}
+
+@media screen and (max-width: map-get($map: $screen-breakpoint, $key: xs)) {
+  .desktopLogo{
+    display: none;
+  }
+
+  .socialBottom {
+    width: 70%;
+
+    svg {
+      width: 1.1rem;
+    }
+  }
+
+  .blogIcons {
+    a {
+      margin: 0.4rem;
+    }
+    width: 100%;
+  }
+
+  .footerMobileLogo {
+    width: 65%;
+  }
+
+  .selectedLink {
+    min-width:auto
+  }
+
+  .logo, .optionContainer{
+    margin-left: 0.5rem;
+  }
+}
+
+.socialBottom {
+  display: flex;
+  flex-direction: row;
+  position: static;
+  align-items: center;
+  width: 100%;
+  z-index: 0;
+  justify-content: end;
+}
+
+.blogIcons {
+  a {
+    display: flex;
+    align-items: center;
+    margin: 0.9rem;
+  }
+  margin-top: 1rem;
+}
+
+.blogIcons a:hover {
+  filter: brightness(1.5);
 }
