+++ src/components/header-banner/header-banner.module.scss
@@ -78,6 +78,7 @@
 
 .dupContent {
   position: absolute;
+  cursor: default;
 }
 
 @keyframes textMove {
