--- src/components/mobile-menu/index.js
@@ -1,1 +1,0 @@
-export { MobileMenu } from "./mobile-menu";
