+++ src/components/nav-menu/mobile-nav-menu.module.scss
@@ -1,5 +1,5 @@
-@import "../../css//colors.scss";
-@import "../../css/variables.scss";
+@import "src/css/colors.scss";
+@import "src/css/variables.scss";
 
 // General Container
