+++ src/components/nav-menu/nav-menu-elements.module.scss
@@ -1,0 +1,19 @@
+@import "/src/css/colors.scss";
+@import "/src/css/variables.scss";
+
+.linkTextContainer {
+  margin: 10px 0 10px 50px;
+  display: inline-block;
+  text-decoration: none;
+}
+
+.link {
+  font-family: $link-font;
+  color: $blueOne;
+  font-size: 16px;
+  line-height: 20px;
+}
+
+.link:hover {
+  color: $grayTwo;
+}
