+++ src/components/nav-menu/nav-menu.module.scss
@@ -1,19 +1,22 @@
 @import "/src/css/colors.scss";
 @import "/src/css/variables.scss";
 
-.linkTextContainer {
-  margin: 10px 0 10px 50px;
-  display: inline-block;
-  text-decoration: none;
+.navMenuContainer {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
 }
 
-.link {
+.getStarted {
+  position: relative;
+  top: 0;
+  transition: top ease 0.2s;
   font-family: $link-font;
-  color: $blueOne;
-  font-size: 16px;
-  line-height: 20px;
+  border: none;
+  cursor: pointer;
 }
 
-.link:hover {
-  color: $grayTwo;
+.getStarted:hover {
+  top: -4px;
+  background-color: $xlPinkThree;
 }
