--- src/elements/form.js
@@ -1,4 +1,0 @@
-import styled from 'styled-components';   
-
-export const StyledForm = styled.form`
-`
