--- src/layouts/index.js
@@ -1,1 +1,0 @@
-export { Layout } from "./layout";
