--- src/layouts/layout.module.scss
@@ -1,143 +1,0 @@
-@import "../css/colors.scss";
-@import "../css/variables.scss";
-
-body {
-  font-family: $primary-font, serif;
-}
-
-.navMenuContainer {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-}
-
-.getStarted {
-  position: relative;
-  top: 0;
-  transition: top ease 0.2s;
-  font-family: $link-font;
-  border: none;
-  cursor: pointer;
-}
-
-.getStarted:hover {
-  top: -4px;
-  background-color: $xlPinkThree;
-}
-
-.desktopLogo {
-  width: 20%;
-}
-
-.mobileBanner {
-  background-color: $blueSix;
-}
-
-.socialBottom {
-  display: flex;
-  flex-direction: row;
-  position: static;
-  align-items: center;
-  width: 100%;
-  z-index: 0;
-  justify-content: end;
-}
-
-.blogIcons {
-  a {
-    display: flex;
-    align-items: center;
-    margin: 0.9rem;
-  }
-  margin-top: 1rem;
-}
-
-.blogIcons a:hover {
-  filter: brightness(1.5);
-}
-
-.logoMobile {
-  display: none;
-}
-
-// Footer
-
-.container {
-  display: flex;
-  flex-direction: column;
-  align-items: start;
-  justify-content: center;
-}
-
-.linkContainer {
-  margin-top: 2rem;
-  display: flex;
-  flex-direction: column;
-  color: $defaultGray;
-}
-
-.optionContainer {
-  margin-left: 0.9rem;
-  display: flex;
-  flex-direction: column;
-  align-items: start;
-  justify-content: center;
-
-  a {
-    cursor: pointer;
-    font-family: $primary-font;
-  }
-
-  a:hover {
-    color: $neutral100;
-  }
-}
-
-.logo {
-  margin-left: 0.9rem;
-}
-
-// Media
-
-@media screen and (max-height: map-get($map: $screen-breakpoint, $key: md)) {
-  .filterContainer {
-    flex-wrap: wrap;
-  }
-
-  .filterElement {
-    margin-bottom: 2rem;
-  }
-}
-
-@media screen and (max-width: map-get($map: $screen-breakpoint, $key: xs)) {
-  .desktopLogo{
-    display: none;
-  }
-
-  .socialBottom {
-    width: 70%;
-
-    svg {
-      width: 1.1rem;
-    }
-  }
-
-  .blogIcons {
-    a {
-      margin: 0.4rem;
-    }
-    width: 100%;
-  }
-
-  .footerMobileLogo {
-    width: 65%;
-  }
-
-  .selectedLink {
-    min-width:auto
-  }
-
-  .logo, .optionContainer{
-    margin-left: 0.5rem;
-  }
-}
